First attempt and ReceiveSplitSend
Not working. Followed Guides. But still not grasping basic concepts of using the successfully deployed smart contract. I.e. You can deploy it, but using the functions of the contract is different to how your imagining now. When coming back, test in Remix again to see what I'm talking about.